IM-37, IM-38 Created ItemDAOTest and ItemTest, updated sql-schema and sql-data to perform item database tests.

+++ src/main/resources/sql-schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS `ims`.`customers` (
 CREATE TABLE IF NOT EXISTS `ims`.`items` (
     `id` INT(11) NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(40) not NULL,
-    `value` float not NULL,
+    `value` Double not NULL,
     PRIMARY KEY (`id`)
 );
 CREATE TABLE IF NOT EXISTS `ims`.`orders` (
